Remove duplicated types
