+++ pkg/amber/provenance.go
@@ -26,8 +26,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/project-oak/transparent-release/pkg/intoto"
 	slsa "github.com/project-oak/transparent-release/pkg/intoto/slsa_provenance/v0.2"
+	"github.com/project-oak/transparent-release/pkg/types"
 	"github.com/xeipuuv/gojsonschema"
 
 	_ "embed"
@@ -46,73 +46,6 @@ var schema []byte
 type BuildConfig struct {
 	Command    []string `json:"command"`
 	OutputPath string   `json:"outputPath"`
-}
-
-// ValidatedProvenance wraps an intoto.Statement representing a valid SLSA provenance statement.
-// A provenance statement is valid if it contains a single subject, with a SHA256 hash.
-// Deprecated: Instead use ValidatedProvenance in `pkg/intoto/slsa_provenance/v0.2`.
-type ValidatedProvenance struct {
-	// The field is private so that invalid instances cannot be created.
-	provenance intoto.Statement
-}
-
-// GetProvenance returns a partial copy of the provenance statement wrapped in this instance.
-// The partial copy guarantees that the validity condition will not be violated.
-func (p *ValidatedProvenance) GetProvenance() intoto.Statement {
-	subject := intoto.Subject{
-		Name:   p.provenance.Subject[0].Name,
-		Digest: intoto.DigestSet{"sha256": p.provenance.Subject[0].Digest["sha256"]},
-	}
-
-	statementHeader := intoto.StatementHeader{
-		Type:          p.provenance.Type,
-		PredicateType: p.provenance.PredicateType,
-		Subject:       []intoto.Subject{subject},
-	}
-
-	return intoto.Statement{
-		StatementHeader: statementHeader,
-		Predicate:       p.provenance.Predicate,
-	}
-}
-
-// GetBinarySHA256Digest returns the SHA256 digest of the subject.
-func (p *ValidatedProvenance) GetBinarySHA256Digest() string {
-	return p.provenance.Subject[0].Digest["sha256"]
-}
-
-// GetBuildType returns the build type of the provenance.
-func (p *ValidatedProvenance) GetBuildType() string {
-	return p.provenance.Predicate.(slsa.ProvenancePredicate).BuildType
-}
-
-// GetBinaryName returns the name of the subject.
-func (p *ValidatedProvenance) GetBinaryName() string {
-	return p.provenance.Subject[0].Name
-}
-
-// GetBuildCmd returns the build command.
-func (p *ValidatedProvenance) GetBuildCmd() ([]string, error) {
-	buildConfig, err := parseBuildConfig(p.provenance.Predicate.(slsa.ProvenancePredicate))
-	if err != nil {
-		return nil, fmt.Errorf("could not parse BuildConfig")
-	}
-	return buildConfig.Command, nil
-}
-
-// GetBuilderImageDigest returns the digest for the Builder Image.
-func (p *ValidatedProvenance) GetBuilderImageDigest() (string, error) {
-	materials := p.provenance.Predicate.([]slsa.ProvenanceMaterial)
-	for _, material := range materials {
-		// This is a crude way to estimate if one of the materials is the builder image.
-		// However, even if we get a "wrong" digest as the builder image, the reference values should
-		// not contain this wrong digest, so worst case verifying the provenance fails, when it should not.
-		if strings.Contains(material.URI, "@sha256:") {
-			digest := material.Digest["sha256"]
-			return digest, nil
-		}
-	}
-	return "", fmt.Errorf("could not find the builder image in %v", materials)
 }
 
 func validateSLSAProvenanceJSON(provenanceFile []byte) error {
@@ -140,58 +73,21 @@ func validateSLSAProvenanceJSON(provenanceFile []byte) error {
 // ParseProvenanceFile reads a JSON file from a given path, and calls ParseProvenanceData on the
 // content of the file, if the read is successful.
 // Returns an error if the file is not a valid provenance statement.
-func ParseProvenanceFile(path string) (*ValidatedProvenance, error) {
+func ParseProvenanceFile(path string) (*types.ValidatedProvenance, error) {
 	statementBytes, err := os.ReadFile(path)
 	if err != nil {
 		return nil, fmt.Errorf("could not read the provenance file: %v", err)
 	}
 
-	return ParseProvenanceData(statementBytes)
-}
-
-// ParseProvenanceData validates the given bytes against the Amber buildType schema, and parses it
-// into an instance of intoto.Statement.
-// Returns an error if the bytes do not represent a valid JSON-encoded provenance statement.
-// Deprecated: Instead use ParseProvenanceData in `pkg/intoto/slsa_provenance/v0.2`.
-func ParseProvenanceData(statementBytes []byte) (*ValidatedProvenance, error) {
 	if err := validateSLSAProvenanceJSON(statementBytes); err != nil {
 		return nil, err
 	}
 
-	var statement intoto.Statement
-	if err := json.Unmarshal(statementBytes, &statement); err != nil {
-		return nil, fmt.Errorf("could not unmarshal the provenance file:\n%v", err)
-	}
-
-	if len(statement.Subject) != 1 || statement.Subject[0].Digest["sha256"] == "" {
-		return nil, fmt.Errorf("the provenance must have exactly one subject with a sha256 digest")
-	}
-
-	// statement.Predicate is now just a map, we have to parse it into an instance of slsa.ProvenancePredicate
-	predicateBytes, err := json.Marshal(statement.Predicate)
-	if err != nil {
-		return nil, fmt.Errorf("could not marshal Predicate map into JSON bytes: %v", err)
-	}
-
-	var predicate slsa.ProvenancePredicate
-	if err = json.Unmarshal(predicateBytes, &predicate); err != nil {
-		return nil, fmt.Errorf("could not unmarshal JSON bytes into a slsa.ProvenancePredicate: %v", err)
-	}
-
-	buildConfig, err := parseBuildConfig(predicate)
-	if err != nil {
-		return nil, fmt.Errorf("could not parse BuildConfig")
-	}
-
-	// Replace maps with objects
-	predicate.BuildConfig = buildConfig
-	statement.Predicate = predicate
-
-	return &ValidatedProvenance{provenance: statement}, nil
+	return types.ParseProvenanceData(statementBytes)
 }
 
-// parseBuildConfig parses the map in predicate.BuildConfig into an instance of BuildConfig.
-func parseBuildConfig(predicate slsa.ProvenancePredicate) (BuildConfig, error) {
+// ParseBuildConfig parses the map in predicate.BuildConfig into an instance of BuildConfig.
+func ParseBuildConfig(predicate slsa.ProvenancePredicate) (BuildConfig, error) {
 	var buildConfig BuildConfig
 	buildConfigBytes, err := json.Marshal(predicate.BuildConfig)
 	if err != nil {
@@ -201,4 +97,28 @@ func parseBuildConfig(predicate slsa.ProvenancePredicate) (BuildConfig, error) {
 		return buildConfig, fmt.Errorf("could not unmarshal JSON bytes into a BuildConfig: %v", err)
 	}
 	return buildConfig, nil
+}
+
+// GetBuildCmd returns the build command.
+func GetBuildCmd(p *types.ValidatedProvenance) ([]string, error) {
+	buildConfig, err := ParseBuildConfig(p.GetProvenance().Predicate.(slsa.ProvenancePredicate))
+	if err != nil {
+		return nil, fmt.Errorf("could not parse BuildConfig")
+	}
+	return buildConfig.Command, nil
+}
+
+// GetBuilderImageDigest returns the digest for the Builder Image.
+func GetBuilderImageDigest(p *types.ValidatedProvenance) (string, error) {
+	materials := p.GetProvenance().Predicate.([]slsa.ProvenanceMaterial)
+	for _, material := range materials {
+		// This is a crude way to estimate if one of the materials is the builder image.
+		// However, even if we get a "wrong" digest as the builder image, the reference values should
+		// not contain this wrong digest, so worst case verifying the provenance fails, when it should not.
+		if strings.Contains(material.URI, "@sha256:") {
+			digest := material.Digest["sha256"]
+			return digest, nil
+		}
+	}
+	return "", fmt.Errorf("could not find the builder image in %v", materials)
 }
