+++ pkg/intoto/slsa_provenance/v0.2/provenance.go
@@ -19,8 +19,6 @@
 package v02
 
 import (
-	"encoding/json"
-	"fmt"
 	"time"
 
 	"github.com/project-oak/transparent-release/pkg/intoto"
@@ -165,76 +163,4 @@ type ProvenanceComplete struct {
 	// Materials if true, means the builder claims that materials is complete, usually through some
 	// controls to prevent network access. Sometimes called “hermetic”.
 	Materials bool `json:"materials"`
-}
-
-// ValidatedProvenance wraps an intoto.Statement representing a valid SLSA provenance statement.
-// A provenance statement is valid if it contains a single subject, with a SHA256 hash.
-type ValidatedProvenance struct {
-	// The field is private so that invalid instances cannot be created.
-	provenance intoto.Statement
-}
-
-// GetBinarySHA256Digest returns the SHA256 digest of the subject.
-func (p *ValidatedProvenance) GetBinarySHA256Digest() string {
-	return p.provenance.Subject[0].Digest["sha256"]
-}
-
-// GetBinaryName returns the name of the subject.
-func (p *ValidatedProvenance) GetBinaryName() string {
-	return p.provenance.Subject[0].Name
-}
-
-// GetBuildType returns the build type of the provenance.
-func (p *ValidatedProvenance) GetBuildType() string {
-	return p.provenance.Predicate.(ProvenancePredicate).BuildType
-}
-
-// GetProvenance returns a partial copy of the provenance statement wrapped in this instance.
-// The partial copy guarantees that the validity condition will not be violated.
-func (p *ValidatedProvenance) GetProvenance() intoto.Statement {
-	subject := intoto.Subject{
-		Name:   p.provenance.Subject[0].Name,
-		Digest: intoto.DigestSet{"sha256": p.provenance.Subject[0].Digest["sha256"]},
-	}
-
-	statementHeader := intoto.StatementHeader{
-		Type:          p.provenance.Type,
-		PredicateType: p.provenance.PredicateType,
-		Subject:       []intoto.Subject{subject},
-	}
-
-	return intoto.Statement{
-		StatementHeader: statementHeader,
-		Predicate:       p.provenance.Predicate,
-	}
-}
-
-// ParseProvenanceData validates that the given bytes represent a valid SLSA provenance.
-// Returns an error if the bytes do not represent a valid JSON-encoded provenance statement.
-// Otherwise returns an instance of ValidatedProvenance.
-func ParseProvenanceData(statementBytes []byte) (*ValidatedProvenance, error) {
-	var statement intoto.Statement
-	if err := json.Unmarshal(statementBytes, &statement); err != nil {
-		return nil, fmt.Errorf("could not unmarshal the provenance file:\n%v", err)
-	}
-
-	if len(statement.Subject) != 1 || statement.Subject[0].Digest["sha256"] == "" {
-		return nil, fmt.Errorf("the provenance must have exactly one subject with a sha256 digest")
-	}
-
-	// statement.Predicate is now just a map, we have to parse it into an instance of slsa.ProvenancePredicate
-	predicateBytes, err := json.Marshal(statement.Predicate)
-	if err != nil {
-		return nil, fmt.Errorf("could not marshal Predicate map into JSON bytes: %v", err)
-	}
-
-	var predicate ProvenancePredicate
-	if err = json.Unmarshal(predicateBytes, &predicate); err != nil {
-		return nil, fmt.Errorf("could not unmarshal JSON bytes into a slsa.ProvenancePredicate: %v", err)
-	}
-
-	// Replace maps with objects
-	statement.Predicate = predicate
-
-	return &ValidatedProvenance{provenance: statement}, nil
 }
